Disable ARCHAPPL_ALL_APPS_ON_ONE_JVM
The variable is only checked to exist, not for its value.
So in order to be 'disabled', it would have to be unset:
https://github.com/slacmshankar/epicsarchiverap/blob/2e4dc4f/src/main/org/epics/archiverappliance/config/DefaultConfigService.java#L439

An environment variable set once in a Docker container cannot easily
be unset at a later point in time, so we omit it in the Dockerfile.
We can still specify it later in our `docker run` command  with the
`-e` flag: `-e ARCHAPPL_ALL_APPS_ON_ONE_JVM=true` or in the
`environment:` section of a docker-compose.yml file.